Solution task 2. Vehicle Extension 100/100

+++ 03. C# Advanced/02. C# OOP/04. Polymorphism/Homework_Polymorphism/02.VehicleExtension/Factories/VehicleFactory.cs
@@ -5,7 +5,7 @@ namespace _02.VehicleExtension.Factories
 {
     public class VehicleFactory
     {
-        public Vehicle ProduceVehicle(string type, double fuelQty, double fuelConsumption, double tankCapacity)
+        public Vehicle ProduceVehicle(string type, double fuelQty, double fuelConsumption, int tankCapacity)
         {
             Vehicle vehicle = null;
