+++ 03. C# Advanced/02. C# OOP/04. Polymorphism/Homework_Polymorphism/02.VehicleExtension/Interfaces/IVehicle.cs
@@ -4,7 +4,7 @@
     {
         double FuelQuantity { get; }
         double FuelConsumption { get; }
-        double TankCapacity { get; }
+        int TankCapacity { get; }
         string Drive(double distance);
         void Refuel(double litres);
     }
